Move plotted variables adornment and variable legend file into shared-variables.
