+++ src/plugins/shared-variables/graph/plotted-variables-adornment/plotted-variables-adornment-registration.tsx
@@ -1,9 +1,9 @@
-import { registerMultiLegendPart } from "../../../components/legend/legend-registration";
+import { registerAdornmentComponentInfo } from "../../../graph/adornments/adornment-component-info";
+import { registerAdornmentContentInfo } from "../../../graph/adornments/adornment-content-info";
+import { registerMultiLegendPart } from "../../../graph/components/legend/legend-registration";
 import {
   getVariableFunctionLegendIdList, heightOfVariableFunctionLegend, VariableFunctionLegend, variableFunctionLegendType
-} from "../../../components/legend/variable-function-legend";
-import { registerAdornmentComponentInfo } from "../../adornment-component-info";
-import { registerAdornmentContentInfo } from "../../adornment-content-info";
+} from "../legend/variable-function-legend";
 import { PlottedVariablesAdornmentModel } from "./plotted-variables-adornment-model";
 import {
   kPlottedVariablesClass, kPlottedVariablesLabelKey, kPlottedVariablesPrefix, kPlottedVariablesType
