+++ src/plugins/shared-variables/graph/plotted-variables-adornment/plotted-variables-adornment-types.ts
